download_sdn/requirements.txt
functions-framework>=3.4.0
google-cloud-storage>=2.14.0
requests>=2.31.0
google-api-python-client>=2.100.0
